modules/curve: fix function casts
The return value of a function should only be casted to a compatible
type; lots of "(int) floor(x)" can be simply changed to the implicit
casting rules.

+++ modules/src/curve/disc.c
@@ -33,7 +33,7 @@ Curve disc(float r, float *i)
   next = &(c->first);
   prev = NULL;
 
-  ir = (int)ceil(r);
+  ir = ceil(r);
   r2 = r*r;
   i2 = (i?*i*(*i):-1.);
